refactor(chat): split GrueneratorModelAdapter into folder modules
Break the 1347-line adapter into a folder:
- types.ts (metadata/config/callbacks + internal part types)
- parseSSEStream.ts (the SSE event generator + progress/result builders)
- truncation.ts (truncateAttachmentContext, pure)
- buildRequestBody.ts (mode-aware payload assembly, pure; shared chat/eigener body)
- index.ts (createGrueneratorModelAdapter factory + run orchestration)

Public API (createGrueneratorModelAdapter + 3 types) re-exported unchanged.
Adds adapterUtils.test.ts (tsx-style, matching the package convention) for the
pure utils. No behavior change.
