--- packages/chat/src/runtime/GrueneratorModelAdapter.ts
@@ -1,1347 +1,0 @@
-import type {
-  ChatModelAdapter,
-  ChatModelRunOptions,
-  ChatModelRunResult,
-  CompleteAttachment,
-} from '@assistant-ui/react';
-import { useChatConfigStore } from '../stores/chatConfigStore';
-import type {
-  ProgressStage,
-  SearchIntent,
-  GeneratedImage,
-  ChatProgress,
-  Citation,
-  FallbackInfo,
-  SearchResult,
-  StreamMetadata,
-  ProgressStep,
-} from '../hooks/useChatGraphStream';
-import { useAgentStore, type ToolKey, type ThreadMode, type SearchMode } from '../stores/chatStore';
-import { getSystemAgent } from '@gruenerator/shared/agents';
-import { triggerDocEditSchema } from '@gruenerator/contracts';
-import { parseAllMentions } from '../lib/mentionParser';
-import { parseSSELine } from '../lib/sseParser';
-import { INTENT_TO_TOOL, DEEP_TOOL_MAP } from '../lib/toolMappings';
-import { pickStageLabels } from '../lib/progressLabels';
-import { useDocumentChatStore } from '../stores/documentChatStore';
-import { streamErrorMessage } from './streamErrorMessage';
-import type { ConfirmActionData, DocumentCreatedData } from '../types/messageMetadata';
-
-export type GrueneratorMessageMetadata = {
-  progress?: ChatProgress;
-  searchResults?: SearchResult[];
-  citations?: Citation[];
-  generatedImage?: GeneratedImage;
-  sharepicData?: import('../hooks/useChatGraphStream').SharepicData;
-  streamMetadata?: StreamMetadata;
-  threadId?: string;
-  followUpSuggestions?: string[];
-  agentId?: string;
-  agentMention?: string;
-  confirmAction?: ConfirmActionData;
-  createdDocument?: DocumentCreatedData;
-  [key: string]: unknown;
-};
-
-export interface GrueneratorAdapterConfig {
-  agentId: string | null;
-  modelId: string;
-  enabledTools: Record<ToolKey, boolean>;
-  threadId: string | null;
-  selectedNotebookId?: string;
-  threadMode?: ThreadMode;
-  searchMode?: SearchMode;
-  customSystemPrompt?: string | null;
-  customRoleName?: string | null;
-  customEnabledTools?: Record<string, boolean> | null;
-  /** Mention key of the active /skill (e.g. 'instagram'). Server appends the
-   *  skill's `skillSystemPrompt` to the agent's systemRole when set. */
-  activeSkillMention?: string | null;
-}
-
-export interface GrueneratorAdapterCallbacks {
-  onThreadCreated?: (threadId: string) => void;
-  onComplete?: (metadata: StreamMetadata) => void;
-}
-
-interface ToolCallPart {
-  type: 'tool-call';
-  toolCallId: string;
-  toolName: string;
-  argsText: string;
-  args: Record<string, string | number | boolean | null>;
-  result?: unknown;
-}
-
-interface SourcePart {
-  type: 'source';
-  sourceType: 'url';
-  id: string;
-  url: string;
-  title?: string;
-  parentId?: string;
-}
-
-interface StreamOutcome {
-  interrupted: boolean;
-  lastResult?: ChatModelRunResult;
-  indexedDocumentIds: string[];
-}
-
-async function* parseSSEStream(
-  response: Response,
-  callbacks: GrueneratorAdapterCallbacks,
-  outcome: StreamOutcome,
-  agentInfo?: { agentId: string; agentMention?: string }
-): AsyncGenerator<ChatModelRunResult, void> {
-  const reader = response.body?.getReader();
-
-  // React Native's fetch doesn't support ReadableStream — read full text as fallback
-  const fullText = !reader ? await response.text() : null;
-
-  const decoder = new TextDecoder();
-  let buffer = '';
-  const currentEvent = { type: '' };
-  let accumulatedText = '';
-  let accumulatedReasoning = '';
-  // Themed progress labels — picked once per turn, stable for the whole stream.
-  const stageLabels = pickStageLabels();
-  let currentProgress: ChatProgress = {
-    stage: 'classifying',
-    message: stageLabels.classifying,
-  };
-  const progressSteps: ProgressStep[] = [
-    { stage: 'classifying', label: stageLabels.classifying, status: 'in-progress' },
-  ];
-
-  function transitionStep(newStage: ProgressStage, labelOverride?: string) {
-    // Mark current in-progress step as completed
-    for (const step of progressSteps) {
-      if (step.status === 'in-progress') {
-        step.status = 'completed';
-        step.completedAt = Date.now();
-      }
-    }
-    // Add new step if it has a label and isn't 'complete'/'error'/'idle'
-    const label = labelOverride || stageLabels[newStage];
-    if (label && newStage !== 'complete' && newStage !== 'error' && newStage !== 'idle') {
-      // Don't duplicate if the same stage already exists
-      if (!progressSteps.some((s) => s.stage === newStage)) {
-        progressSteps.push({ stage: newStage, label, status: 'in-progress' });
-      } else {
-        // Re-activate existing step
-        const existing = progressSteps.find((s) => s.stage === newStage);
-        if (existing) existing.status = 'in-progress';
-      }
-    }
-    if (newStage === 'complete') {
-      for (const step of progressSteps) {
-        if (step.status !== 'failed') {
-          step.status = 'completed';
-          step.completedAt ??= Date.now();
-        }
-      }
-    }
-    if (newStage === 'error') {
-      for (const step of progressSteps) {
-        if (step.status === 'in-progress') {
-          step.status = 'failed';
-        }
-      }
-    }
-  }
-
-  let receivedSearchResults: SearchResult[] = [];
-  let receivedCitations: Citation[] = [];
-  let receivedImage: GeneratedImage | null = null;
-  let receivedSharepicData: import('../hooks/useChatGraphStream').SharepicData | null = null;
-  let receivedFollowUpSuggestions: string[] = [];
-  let receivedMetadata: StreamMetadata | null = null;
-  let receivedConfirmAction: ConfirmActionData | null = null;
-  let receivedCreatedDocument: DocumentCreatedData | null = null;
-  let activeToolCall: ToolCallPart | null = null;
-  const allToolCalls: ToolCallPart[] = [];
-  let interruptPending = false;
-  let lastYieldTime = 0;
-  const YIELD_INTERVAL = 50; // ms — max 20 yields/sec, matches NotebookModelAdapter
-
-  function buildResult(): ChatModelRunResult {
-    const content: Array<
-      | { type: 'text'; text: string }
-      | { type: 'reasoning'; text: string }
-      | ToolCallPart
-      | SourcePart
-    > = [];
-
-    const groupId = activeToolCall ? activeToolCall.toolCallId : allToolCalls[0]?.toolCallId;
-
-    if (accumulatedReasoning) {
-      content.push({ type: 'reasoning' as const, text: accumulatedReasoning });
-    }
-
-    for (const tc of allToolCalls) {
-      content.push(tc);
-    }
-    if (activeToolCall && !allToolCalls.includes(activeToolCall)) {
-      content.push(activeToolCall);
-    }
-
-    for (const citation of receivedCitations) {
-      if (citation.url) {
-        content.push({
-          type: 'source' as const,
-          sourceType: 'url' as const,
-          id: `source-${citation.id}`,
-          url: citation.url,
-          title: citation.title || undefined,
-          parentId: groupId,
-        });
-      }
-    }
-
-    // Normalize [cite:N] → [N] before emitting, so CitationMarkdownText's
-    // placeholder-badge layer (citationProcessing.ts:25) renders inline-reserved
-    // boxes during streaming. Mirrors NotebookModelAdapter's buildResult and
-    // prevents the SearchGraph "[cite:N]" markers from appearing as plain text.
-    content.push({
-      type: 'text' as const,
-      text: accumulatedText.replace(/\[cite:(\d+)\]/g, '[$1]'),
-    });
-
-    const custom: GrueneratorMessageMetadata = {
-      progress: { ...currentProgress, steps: [...progressSteps] },
-    };
-    if (receivedSearchResults.length > 0) custom.searchResults = receivedSearchResults;
-    if (receivedCitations.length > 0) custom.citations = receivedCitations;
-    if (receivedImage) custom.generatedImage = receivedImage;
-    if (receivedSharepicData) custom.sharepicData = receivedSharepicData;
-    if (receivedMetadata) custom.streamMetadata = receivedMetadata;
-    if (receivedFollowUpSuggestions.length > 0)
-      custom.followUpSuggestions = receivedFollowUpSuggestions;
-    if (receivedConfirmAction) custom.confirmAction = receivedConfirmAction;
-    if (receivedCreatedDocument) custom.createdDocument = receivedCreatedDocument;
-    if (agentInfo?.agentId) {
-      custom.agentId = agentInfo.agentId;
-      if (agentInfo.agentMention) custom.agentMention = agentInfo.agentMention;
-    }
-
-    const isInterrupted = interruptPending && currentProgress.stage === 'complete';
-
-    return {
-      content,
-      metadata: { custom },
-      ...(isInterrupted
-        ? { status: { type: 'requires-action' as const, reason: 'tool-calls' as const } }
-        : {}),
-    };
-  }
-
-  // Feed chunks: stream from reader, or single chunk from full text fallback
-  const chunks: AsyncIterable<Uint8Array> = reader
-    ? {
-        [Symbol.asyncIterator]: () => ({
-          next: async () => {
-            const { done, value } = await reader.read();
-            return done
-              ? { done: true as const, value: undefined }
-              : { done: false as const, value };
-          },
-        }),
-      }
-    : (async function* () {
-        yield new TextEncoder().encode(fullText!);
-      })();
-
-  for await (const value of chunks) {
-    buffer += decoder.decode(value, { stream: true });
-    const lines = buffer.split('\n');
-    buffer = lines.pop() || '';
-
-    for (const line of lines) {
-      const { event, data } = parseSSELine(line, currentEvent);
-      if (!event || !data) continue;
-
-      switch (event) {
-        case 'thread_created': {
-          const { threadId: tid } = data as { threadId: string };
-          callbacks.onThreadCreated?.(tid);
-          // Backend has now persisted any seeded initialAssistantMessage as
-          // the first row of this thread. Drop the local copy so a future
-          // new-thread creation doesn't replay a stale seed.
-          useAgentStore.getState().setPendingInitialAssistantMessage(null);
-          break;
-        }
-
-        case 'intent': {
-          const { intent, message, reasoning, searchQuery, subQueries, searchSources } = data as {
-            intent: SearchIntent;
-            message: string;
-            reasoning?: string;
-            searchQuery?: string;
-            subQueries?: string[] | null;
-            searchSources?: string[] | null;
-          };
-          let stage: ProgressStage = 'searching';
-          if (intent === 'direct') stage = 'generating';
-          else if (intent === 'image' || intent === 'sharepic') stage = 'generating_image';
-          else if (intent === 'summary') stage = 'summarizing';
-          transitionStep(stage);
-          currentProgress = { stage, message, intent, reasoning };
-
-          const toolName = INTENT_TO_TOOL[intent];
-          if (toolName) {
-            const hasMultiSearch =
-              (subQueries && subQueries.length > 0) || (searchSources && searchSources.length > 1);
-
-            if (hasMultiSearch) {
-              const queries = subQueries?.length ? subQueries : [searchQuery ?? ''];
-              const sources =
-                searchSources?.length && searchSources.length > 1 ? searchSources : [null];
-
-              for (let i = 0; i < queries.length; i++) {
-                for (const src of sources) {
-                  const effToolName =
-                    src === 'web'
-                      ? 'web_search'
-                      : src === 'documents'
-                        ? 'gruenerator_search'
-                        : toolName;
-                  allToolCalls.push({
-                    type: 'tool-call',
-                    toolCallId: `tc_${Date.now()}_${i}_${src || 'default'}`,
-                    toolName: effToolName,
-                    args: { query: queries[i] },
-                    argsText: JSON.stringify({ query: queries[i] }),
-                  });
-                }
-              }
-              activeToolCall = null;
-            } else {
-              // Don't fall back to `message` (the German status copy) — that
-              // would display the status as the user's query. Empty string lets
-              // the UI hide the chip text gracefully.
-              const toolArgs = { query: searchQuery ?? '' };
-              activeToolCall = {
-                type: 'tool-call',
-                toolCallId: `tc_${Date.now()}`,
-                toolName,
-                args: toolArgs,
-                argsText: JSON.stringify(toolArgs),
-              };
-            }
-          }
-          yield buildResult();
-          break;
-        }
-
-        case 'search_start': {
-          const { message } = data as { message: string };
-          transitionStep('searching');
-          currentProgress = { ...currentProgress, stage: 'searching', message };
-          yield buildResult();
-          break;
-        }
-
-        case 'search_complete': {
-          const { message, resultCount, results, researchMeta, examplesResult } = data as {
-            message: string;
-            resultCount: number;
-            results?: SearchResult[];
-            researchMeta?: unknown;
-            examplesResult?: { press?: unknown[]; social?: unknown[]; message?: string };
-          };
-          if (results) receivedSearchResults = results;
-          // Update searching step label with result count
-          const searchStep = progressSteps.find((s) => s.stage === 'searching');
-          if (searchStep) {
-            searchStep.label = `${resultCount} Ergebnisse`;
-            searchStep.status = 'completed';
-            searchStep.completedAt = Date.now();
-          }
-          transitionStep('generating');
-          currentProgress = {
-            ...currentProgress,
-            stage: 'generating',
-            message,
-            resultCount,
-          };
-
-          // Pick the right result shape per tool:
-          // - research → researchMeta (rich orchestrator result)
-          // - gruenerator_pressemitteilung_examples → { examples: press[], results }
-          // - gruenerator_examples_search → { examples: social[], results }
-          // - everything else → { results }
-          const resultForTool = (toolName: string) => {
-            if (toolName === 'research' && researchMeta != null) {
-              return researchMeta as Record<string, unknown>;
-            }
-            if (toolName === 'gruenerator_pressemitteilung_examples' && examplesResult?.press) {
-              return { results: results || [], examples: examplesResult.press };
-            }
-            if (toolName === 'gruenerator_examples_search' && examplesResult?.social) {
-              return { results: results || [], examples: examplesResult.social };
-            }
-            return { results: results || [] };
-          };
-
-          for (let i = 0; i < allToolCalls.length; i++) {
-            if (!allToolCalls[i].result) {
-              allToolCalls[i] = {
-                ...allToolCalls[i],
-                result: resultForTool(allToolCalls[i].toolName),
-              };
-            }
-          }
-          if (activeToolCall) {
-            activeToolCall = Object.assign({}, activeToolCall, {
-              result: resultForTool(activeToolCall.toolName),
-            });
-          }
-          yield buildResult();
-          break;
-        }
-
-        case 'summary_start': {
-          const { message } = data as { message: string };
-          transitionStep('summarizing');
-          currentProgress = { ...currentProgress, stage: 'summarizing', message };
-          yield buildResult();
-          break;
-        }
-
-        case 'summary_complete': {
-          const { message } = data as { message: string };
-          transitionStep('generating');
-          currentProgress = { ...currentProgress, stage: 'generating', message };
-          yield buildResult();
-          break;
-        }
-
-        case 'image_start': {
-          const { message } = data as { message: string };
-          transitionStep('generating_image');
-          currentProgress = { ...currentProgress, stage: 'generating_image', message };
-          yield buildResult();
-          break;
-        }
-
-        case 'image_complete': {
-          const {
-            message,
-            image,
-            error: imageError,
-          } = data as {
-            message: string;
-            image?: GeneratedImage;
-            error?: string;
-          };
-          if (image) receivedImage = image;
-          transitionStep(imageError ? 'error' : 'generating');
-          currentProgress = {
-            ...currentProgress,
-            stage: imageError ? 'error' : 'generating',
-            message,
-          };
-          yield buildResult();
-          break;
-        }
-
-        case 'sharepic_complete': {
-          const payload = data as {
-            message: string;
-            variants?: import('../hooks/useChatGraphStream').SharepicVariant[];
-            canvasType?: string;
-            initialProps?: Record<string, unknown>;
-            alternatives?: unknown[];
-            error?: string;
-          };
-          if (!payload.error) {
-            if (payload.variants && payload.variants.length > 0) {
-              receivedSharepicData = { variants: payload.variants };
-            } else if (payload.canvasType && payload.initialProps) {
-              const legacyId =
-                typeof crypto !== 'undefined' && 'randomUUID' in crypto
-                  ? crypto.randomUUID()
-                  : `legacy-${Date.now()}`;
-              receivedSharepicData = {
-                variants: [
-                  {
-                    id: legacyId,
-                    canvasType: payload.canvasType,
-                    initialProps: payload.initialProps,
-                  },
-                ],
-              };
-            }
-          }
-          transitionStep(payload.error ? 'error' : 'generating');
-          currentProgress = {
-            ...currentProgress,
-            stage: payload.error ? 'error' : 'generating',
-            message: payload.message,
-          };
-          yield buildResult();
-          break;
-        }
-
-        case 'response_start': {
-          const { message } = data as { message: string };
-          transitionStep('generating');
-          currentProgress = { ...currentProgress, stage: 'generating', message };
-          yield buildResult();
-          break;
-        }
-
-        case 'thinking_step': {
-          const { stepId, toolName, title, status, args, result } = data as {
-            stepId: string;
-            toolName: string;
-            title: string;
-            status: 'in_progress' | 'completed';
-            args?: Record<string, unknown>;
-            result?: {
-              resultCount?: number;
-              results?: unknown[];
-              image?: unknown;
-              error?: string;
-            };
-          };
-
-          const mappedToolName = DEEP_TOOL_MAP[toolName] || toolName;
-
-          if (status === 'in_progress') {
-            // The backend heartbeat (responseStreamingService.startResponseHeartbeat)
-            // re-emits the SAME stepId every 3s. If it matches the current
-            // activeToolCall, or is already in allToolCalls, skip the
-            // archive-and-replace below — otherwise we'd render two tool-call
-            // parts with the same toolCallId and trip assistant-ui's
-            // `tapResources` with "Duplicate key toolCallId-…".
-            const isDuplicateStepId =
-              (activeToolCall !== null && activeToolCall.toolCallId === stepId) ||
-              allToolCalls.some((tc) => tc.toolCallId === stepId);
-            if (!isDuplicateStepId) {
-              // Preserve any pre-existing activeToolCall (e.g. the intent-derived
-              // gruenerator_examples_search tool-call set by the `intent` event)
-              // before overwriting with this thinking_step's tool. Otherwise the
-              // examples tool-call is orphaned and search_complete's result has
-              // nowhere to land — leaving the UI with classify/rerank chips and
-              // no examples card.
-              if (activeToolCall !== null && !allToolCalls.includes(activeToolCall)) {
-                allToolCalls.push(activeToolCall);
-              }
-              const toolArgs = { query: (args?.query as string) || title, ...args };
-              activeToolCall = {
-                type: 'tool-call',
-                toolCallId: stepId,
-                toolName: mappedToolName,
-                args: toolArgs as Record<string, string | number | boolean | null>,
-                argsText: JSON.stringify(toolArgs),
-              };
-            }
-            currentProgress = { stage: 'searching', message: title };
-          } else if (status === 'completed') {
-            if (activeToolCall?.toolCallId === stepId) {
-              activeToolCall = { ...activeToolCall, result: result || {} };
-              allToolCalls.push(activeToolCall);
-              activeToolCall = null;
-            }
-            currentProgress = { stage: 'generating', message: title };
-          }
-          yield buildResult();
-          break;
-        }
-
-        case 'progress_step': {
-          // Internal pipeline stage (classify, rerank, brief). Updates the
-          // progress indicator but MUST NOT touch activeToolCall/allToolCalls
-          // — those are reserved for user-facing tools dispatched via the
-          // `intent` event + `thinking_step`. Conflating the two is what
-          // caused the search→rerank race that orphaned the rich
-          // examples/search/web tool-cards (see PR history).
-          const { title, status } = data as {
-            stepId: string;
-            toolName: string;
-            title: string;
-            status: 'in_progress' | 'completed';
-          };
-          if (status === 'in_progress') {
-            currentProgress = { ...currentProgress, stage: 'searching', message: title };
-          } else if (status === 'completed') {
-            currentProgress = { ...currentProgress, message: title };
-          }
-          yield buildResult();
-          break;
-        }
-
-        case 'text_delta': {
-          const delta = (data as { text: string }).text;
-          accumulatedText += delta;
-          const now = performance.now();
-          if (now - lastYieldTime >= YIELD_INTERVAL) {
-            lastYieldTime = now;
-            yield buildResult();
-          }
-          break;
-        }
-
-        case 'reasoning_delta': {
-          accumulatedReasoning += (data as { text: string }).text;
-          const now = performance.now();
-          if (now - lastYieldTime >= YIELD_INTERVAL) {
-            lastYieldTime = now;
-            yield buildResult();
-          }
-          break;
-        }
-
-        case 'fallback': {
-          // Server switched models silently — log only, no UI.
-          const info = data as FallbackInfo;
-          console.warn(
-            `[GrueneratorModelAdapter] Model fallback: ${info.from.id} → ${info.to.id} (${info.reason})`
-          );
-          break;
-        }
-
-        case 'interrupt': {
-          interruptPending = true;
-          yield buildResult();
-          break;
-        }
-
-        case 'done': {
-          const {
-            citations: cit,
-            generatedImage: img,
-            metadata,
-            interrupted,
-          } = data as {
-            threadId?: string;
-            citations?: Citation[];
-            generatedImage?: GeneratedImage;
-            metadata?: StreamMetadata;
-            interrupted?: boolean;
-          };
-          if (cit) receivedCitations = cit;
-          if (img) receivedImage = img;
-          if (metadata) receivedMetadata = metadata;
-          if (interrupted) interruptPending = true;
-          transitionStep('complete');
-          currentProgress = { stage: 'complete', message: '' };
-          break;
-        }
-
-        case 'confirm_action': {
-          receivedConfirmAction = data as ConfirmActionData;
-          yield buildResult();
-          break;
-        }
-
-        case 'document_created': {
-          receivedCreatedDocument = data as DocumentCreatedData;
-          yield buildResult();
-          break;
-        }
-
-        case 'document_indexed': {
-          const { documentId } = data as { documentId: string };
-          outcome.indexedDocumentIds.push(documentId);
-          break;
-        }
-
-        case 'trigger_doc_edit': {
-          // Live document edit (docs editor surface). The chat backend has
-          // classified intent=edit_current_doc and forwards the user's prompt
-          // here so the docs frontend can dispatch into BlockNote's AIExtension.
-          // Handlers are keyed by documentId — there's exactly one docs surface
-          // per document, registered when DocsAssistantChat mounts.
-          const parsed = triggerDocEditSchema.safeParse(data);
-          if (!parsed.success) {
-            console.warn('[ChatAdapter] trigger_doc_edit payload failed validation', parsed.error);
-            break;
-          }
-          const payload = parsed.data;
-          const handler = useChatConfigStore
-            .getState()
-            .documentEditHandlers.get(payload.targetDocumentId);
-          if (handler) {
-            try {
-              await handler(payload);
-            } catch (err) {
-              console.warn('[ChatAdapter] documentEditHandler threw', err);
-            }
-          } else {
-            console.warn(
-              '[ChatAdapter] trigger_doc_edit received but no handler registered for doc',
-              payload.targetDocumentId
-            );
-          }
-          break;
-        }
-
-        // ── Search mode events ──
-        case 'sources_preview': {
-          const { results: previewResults, resultCount } = data as {
-            results?: SearchResult[];
-            resultCount?: number;
-          };
-          if (previewResults) {
-            receivedSearchResults = previewResults;
-            // Create synthetic search_sources tool call for sources-first rendering
-            const sourcesToolCall: ToolCallPart = {
-              type: 'tool-call',
-              toolCallId: `tc_sources_${Date.now()}`,
-              toolName: 'search_sources',
-              argsText: '{}',
-              args: {},
-              result: {
-                results: previewResults,
-                resultCount: resultCount || previewResults.length,
-              },
-            };
-            allToolCalls.push(sourcesToolCall);
-          }
-          transitionStep('generating', 'Generierung');
-          break;
-        }
-
-        case 'suggestions': {
-          const { suggestions: sugg } = data as { suggestions?: string[] };
-          if (sugg) {
-            receivedFollowUpSuggestions = sugg;
-          }
-          break;
-        }
-
-        case 'research_step': {
-          const { step, message } = data as { step: string; message: string };
-          transitionStep(step as ProgressStage, message);
-          currentProgress = { stage: step as ProgressStage, message };
-          break;
-        }
-
-        // ── Notebook mode events ──
-        case 'completion': {
-          const completionData = data as {
-            text?: string;
-            citations?: Array<{
-              index: string;
-              cited_text?: string;
-              document_title?: string;
-              document_id?: string;
-              source_url?: string | null;
-              similarity_score?: number;
-              chunk_index?: number;
-              collection_id?: string;
-              collection_name?: string;
-            }>;
-          };
-          if (completionData.text) {
-            // Normalize [cite:N] markers to [N]
-            accumulatedText = completionData.text.replace(/\[cite:(\d+)\]/g, '[$1]');
-          }
-          if (completionData.citations) {
-            receivedCitations = completionData.citations.map((c) => ({
-              id: parseInt(c.index, 10),
-              title: c.document_title ?? '',
-              url: c.source_url ?? '',
-              snippet: c.cited_text ?? '',
-              citedText: c.cited_text,
-              source: c.collection_name ?? '',
-              collectionName: c.collection_name,
-              documentId: c.document_id,
-              chunkIndex: c.chunk_index,
-              similarityScore: c.similarity_score,
-              collectionId: c.collection_id,
-            }));
-          }
-          transitionStep('complete');
-          currentProgress = { stage: 'complete', message: '' };
-          break;
-        }
-
-        case 'error': {
-          const { error } = data as { error: string };
-          transitionStep('error');
-          throw new Error(error);
-        }
-      }
-    }
-  }
-
-  const finalResult = buildResult();
-  outcome.lastResult = finalResult;
-  yield finalResult;
-
-  outcome.interrupted = interruptPending;
-
-  if (receivedMetadata && !interruptPending) {
-    callbacks.onComplete?.(receivedMetadata);
-  }
-}
-
-function truncateAttachmentContext(text: string, maxChars: number): string | undefined {
-  if (!text) return undefined;
-  if (text.length <= maxChars) return text;
-  const half = Math.floor((maxChars - 40) / 2);
-  return `${text.slice(0, half)}\n\n[...gekürzt...]\n\n${text.slice(-half)}`;
-}
-
-export function createGrueneratorModelAdapter(
-  getConfig: () => GrueneratorAdapterConfig,
-  callbacks: GrueneratorAdapterCallbacks
-): ChatModelAdapter {
-  // Tracks which thread has a pending HITL interrupt — persists across run() calls
-  let interruptedThreadId: string | null = null;
-  let lastInterruptedResult: ChatModelRunResult | null = null;
-
-  return {
-    async *run(options: ChatModelRunOptions): AsyncGenerator<ChatModelRunResult, void> {
-      const { messages, abortSignal } = options;
-      const config = getConfig();
-
-      // unstable_getMessage() provides the current assistant message (not in messages array).
-      // This is where addResult() writes the user's answer for human tool calls.
-      const currentAssistant = options.unstable_getMessage?.();
-
-      // Resume detection via unstable_getMessage() — the canonical way to read addResult() answers.
-      // assistant-ui writes the result onto the current assistant message, NOT into messages[].
-      if (currentAssistant) {
-        const askHumanResult = currentAssistant.content?.find(
-          (p) =>
-            p.type === 'tool-call' &&
-            p.toolName === 'ask_human' &&
-            'result' in p &&
-            typeof p.result === 'string' &&
-            (p.result as string).length > 0
-        );
-        if (askHumanResult && askHumanResult.type === 'tool-call') {
-          const answer = String(askHumanResult.result);
-          interruptedThreadId = null;
-          lastInterruptedResult = null;
-          const { fetch: configFetch, endpoints } = useChatConfigStore.getState();
-          const resumeResponse = await configFetch(endpoints.chatResume, {
-            method: 'POST',
-            headers: { 'Content-Type': 'application/json' },
-            body: JSON.stringify({
-              threadId: config.threadId,
-              resume: answer,
-            }),
-            signal: abortSignal,
-          });
-
-          if (!resumeResponse.ok) {
-            const errorData = await resumeResponse.json().catch(() => ({}));
-            throw new Error(
-              (errorData as { error?: string }).error || `HTTP error ${resumeResponse.status}`
-            );
-          }
-
-          const resumeOutcome: StreamOutcome = { interrupted: false, indexedDocumentIds: [] };
-          yield* parseSSEStream(
-            resumeResponse,
-            callbacks,
-            resumeOutcome,
-            config.agentId ? { agentId: config.agentId } : undefined
-          );
-          if (resumeOutcome.interrupted) {
-            interruptedThreadId = config.threadId;
-            lastInterruptedResult = resumeOutcome.lastResult ?? null;
-          }
-          return;
-        }
-
-        // Current assistant has pending ask_human without result — spurious re-invocation.
-        const pendingAskHuman = currentAssistant.content?.find(
-          (p) =>
-            p.type === 'tool-call' && p.toolName === 'ask_human' && !('result' in p && p.result)
-        );
-        if (pendingAskHuman) {
-          console.warn('[ModelAdapter] BLOCKED — pending ask_human without answer');
-          throw new DOMException('Aborted', 'AbortError');
-        }
-      }
-
-      // Stateful guard: block re-invocation if we know this thread has a pending interrupt
-      // (covers case where unstable_getMessage returns undefined after history rehydration)
-      if (interruptedThreadId && interruptedThreadId === config.threadId) {
-        console.warn('[ModelAdapter] BLOCKED — thread has pending interrupt');
-        throw new DOMException('Aborted', 'AbortError');
-      }
-
-      // Clear stale interrupt if switching to a different thread
-      if (interruptedThreadId && interruptedThreadId !== config.threadId) {
-        interruptedThreadId = null;
-        lastInterruptedResult = null;
-      }
-
-      const formattedMessages = messages.map((m) => {
-        const parts: Array<
-          | { type: 'text'; text: string }
-          | { type: 'image'; image: string }
-          | { type: 'file'; name: string; mimeType: string; data: string }
-        > = [];
-
-        for (const part of m.content) {
-          if (part.type === 'text') {
-            parts.push({ type: 'text', text: part.text });
-          } else if (part.type === 'image') {
-            parts.push({ type: 'image', image: part.image });
-          } else if (part.type === 'file') {
-            parts.push({
-              type: 'file',
-              name: (part as { name?: string }).name ?? 'file',
-              mimeType: (part as { mimeType?: string }).mimeType ?? 'application/octet-stream',
-              data: (part as { data?: string }).data ?? '',
-            });
-          }
-        }
-
-        // Merge attachment content parts into formattedMessages so the backend
-        // can also see them when inspecting the messages array directly.
-        if (m.role === 'user' && 'attachments' in m) {
-          const attachments = (m as { attachments: readonly CompleteAttachment[] }).attachments;
-          for (const att of attachments) {
-            for (const part of att.content) {
-              if (part.type === 'text') {
-                parts.push({ type: 'text', text: part.text });
-              } else if (part.type === 'image') {
-                parts.push({ type: 'image', image: part.image });
-              } else if (part.type === 'file') {
-                parts.push({
-                  type: 'file',
-                  name: att.name ?? 'file',
-                  mimeType: (part as { mimeType?: string }).mimeType ?? 'application/octet-stream',
-                  data: (part as { data?: string }).data ?? '',
-                });
-              }
-            }
-          }
-        }
-
-        if (parts.length === 0) {
-          parts.push({ type: 'text', text: '' });
-        }
-
-        return { id: m.id, role: m.role, parts };
-      });
-
-      // Resolve effective mode: an agent with routeTo='search' forces 'search' mode
-      // regardless of the store's threadMode (the agent IS the mode). Legacy
-      // threadMode='search' without a search-agent collapses to 'chat' — the old
-      // Suche toggle was removed, so the only path to SearchGraph is the agent.
-      const activeAgentForRouting = config.agentId ? getSystemAgent(config.agentId) : null;
-      const storedMode = config.threadMode || 'chat';
-      const effectiveMode: ThreadMode =
-        activeAgentForRouting?.routeTo === 'search'
-          ? 'search'
-          : storedMode === 'search'
-            ? 'chat'
-            : storedMode;
-
-      // Surface tools (edit_current_doc) belong to the surface, not the agent —
-      // but if the user picks a search-route agent, SearchGraph can't run them.
-      // Strip the edit hook; keep save_as_doc, which is harmless.
-      const safeCustomEnabledTools =
-        activeAgentForRouting?.routeTo === 'search' && config.customEnabledTools
-          ? Object.fromEntries(
-              Object.entries(config.customEnabledTools).filter(([k]) => k !== 'edit_current_doc')
-            )
-          : config.customEnabledTools;
-
-      // Skip attachment extraction and mention parsing for non-chat modes
-      const isChatMode = effectiveMode === 'chat' || effectiveMode === 'eigener';
-
-      // Extract attachments from AUI's CompleteAttachment objects on the last user message.
-      // AUI stores file/image content in message.attachments[].content, NOT in message.content.
-      const extractedAttachments: Array<{
-        name: string;
-        type: string;
-        size: number;
-        data: string;
-        isImage: boolean;
-      }> = [];
-
-      const lastUserMsg = [...messages].reverse().find((m) => m.role === 'user');
-      if (isChatMode && lastUserMsg && 'attachments' in lastUserMsg) {
-        for (const attachment of lastUserMsg.attachments as unknown as Array<{
-          name?: string;
-          contentType?: string;
-          content: Array<
-            | { type: 'file'; data: string; mimeType: string }
-            | { type: 'image'; image: string }
-            | { type: string }
-          >;
-        }>) {
-          for (const part of attachment.content) {
-            if (part.type === 'file' && 'data' in part) {
-              extractedAttachments.push({
-                name: attachment.name || 'file',
-                type: (part as { mimeType: string }).mimeType,
-                size: Math.ceil(((part as { data: string }).data.length * 3) / 4),
-                data: (part as { data: string }).data,
-                isImage: false,
-              });
-            } else if (part.type === 'image' && 'image' in part) {
-              const imageData = (part as { image: string }).image;
-              const commaIdx = imageData.indexOf(',');
-              const header = commaIdx > 0 ? imageData.slice(0, commaIdx) : '';
-              const base64Data = commaIdx > 0 ? imageData.slice(commaIdx + 1) : imageData;
-              const mimeMatch = header.match(/data:(.*?);/);
-              extractedAttachments.push({
-                name: attachment.name || 'image',
-                type: mimeMatch?.[1] || 'image/jpeg',
-                size: Math.ceil((base64Data.length * 3) / 4),
-                data: base64Data,
-                isImage: true,
-              });
-            }
-          }
-        }
-      }
-
-      // Extract @-mentions from the last user message for agent routing + notebook/document scoping
-      // Only applies in chat mode — search and notebook modes don't use mentions
-      let effectiveAgentId = config.agentId;
-      let effectiveAgentMention: string | undefined;
-      let notebookIds: string[] = [];
-      let forcedTools: string[] = [];
-      let documentIds: string[] = [];
-      let textIds: string[] = [];
-      let boardIds: string[] = [];
-      let docMentionIds: string[] = [];
-      let wolkeFiles: ReturnType<typeof parseAllMentions>['wolkeFiles'] = [];
-      let connectFiles: ReturnType<typeof parseAllMentions>['connectFiles'] = [];
-      let hasDocumentChat = false;
-      if (isChatMode)
-        for (let i = formattedMessages.length - 1; i >= 0; i--) {
-          const msg = formattedMessages[i];
-          if (msg.role !== 'user') continue;
-          const textPart = msg.parts.find(
-            (p): p is { type: 'text'; text: string } => p.type === 'text'
-          );
-          if (textPart) {
-            const parsed = parseAllMentions(textPart.text);
-            // Only override the URL/store-derived agent when the user actually
-            // typed an @agent or /skill mention. parseAllMentions falls back to
-            // getDefaultAgent() (universal) when no mention exists, which would
-            // otherwise silently wipe ?agent=… deep links.
-            if (parsed.agentMention !== undefined) {
-              effectiveAgentId = parsed.agentId;
-              effectiveAgentMention = parsed.agentMention;
-            }
-            notebookIds = parsed.notebookIds;
-            forcedTools = parsed.forcedTools;
-            documentIds = parsed.documentIds;
-            textIds = parsed.textIds;
-            boardIds = parsed.boardIds;
-            docMentionIds = parsed.docMentionIds;
-            wolkeFiles = parsed.wolkeFiles;
-            connectFiles = parsed.connectFiles;
-            hasDocumentChat = parsed.hasDocumentChat;
-            textPart.text = parsed.cleanText;
-
-            if (parsed.unresolvedMentions.length > 0) {
-              const names = parsed.unresolvedMentions.map((m) => `@${m}`).join(', ');
-              console.warn(
-                `[ModelAdapter] Unresolved mentions: ${names} — use @docs to browse collaborative documents`
-              );
-              if (typeof window !== 'undefined') {
-                window.dispatchEvent(
-                  new CustomEvent('gruenerator:unresolved-mentions', {
-                    detail: { mentions: parsed.unresolvedMentions },
-                  })
-                );
-              }
-            }
-          }
-          break;
-        }
-
-      // Extract doc/document mentions from the user message's attachments
-      // (the "tag" UX). Each picked doc was attached as an assistant-ui
-      // CompleteAttachment with a `data` content part carrying our metadata.
-      if (isChatMode && lastUserMsg && 'attachments' in lastUserMsg) {
-        const seenDocs = new Set(documentIds);
-        const seenTexts = new Set(textIds);
-        const seenCollab = new Set(docMentionIds);
-        const seenWolke = new Set(wolkeFiles.map((f) => `${f.shareLinkId}:${f.path}`));
-        const seenConnect = new Set(connectFiles.map((f) => `${f.provider}:${f.fileId}`));
-        type GruenMentionData =
-          | { kind: 'collab'; id: string; slug: string; title: string }
-          | {
-              kind: 'document';
-              documentId: string;
-              sourceType: 'notebook' | 'document' | 'text';
-            }
-          | { kind: 'wolke'; shareLinkId: string; path: string; name: string }
-          | { kind: 'connect'; provider: string; fileId: string; name: string; mimeType?: string };
-        const attachments = (lastUserMsg as { attachments: readonly CompleteAttachment[] })
-          .attachments;
-        for (const att of attachments) {
-          if (!att.contentType?.startsWith('application/x-gruenerator-')) continue;
-          for (const part of att.content) {
-            if (part.type !== 'data') continue;
-            const dataPart = part as { type: 'data'; name?: string; data: GruenMentionData };
-            if (dataPart.name !== 'gruenerator-mention') continue;
-            const data = dataPart.data;
-            if (data.kind === 'collab') {
-              if (!seenCollab.has(data.id)) {
-                seenCollab.add(data.id);
-                docMentionIds.push(data.id);
-              }
-            } else if (data.kind === 'document') {
-              if (data.sourceType === 'text') {
-                if (!seenTexts.has(data.documentId)) {
-                  seenTexts.add(data.documentId);
-                  textIds.push(data.documentId);
-                }
-              } else {
-                if (!seenDocs.has(data.documentId)) {
-                  seenDocs.add(data.documentId);
-                  documentIds.push(data.documentId);
-                }
-              }
-            } else if (data.kind === 'wolke') {
-              const key = `${data.shareLinkId}:${data.path}`;
-              if (!seenWolke.has(key)) {
-                seenWolke.add(key);
-                wolkeFiles.push({
-                  shareLinkId: data.shareLinkId,
-                  path: data.path,
-                  name: data.name,
-                });
-              }
-            } else if (data.kind === 'connect') {
-              const key = `${data.provider}:${data.fileId}`;
-              if (!seenConnect.has(key)) {
-                seenConnect.add(key);
-                connectFiles.push({
-                  provider: data.provider,
-                  fileId: data.fileId,
-                  name: data.name,
-                  ...(data.mimeType ? { mimeType: data.mimeType } : {}),
-                });
-              }
-            }
-          }
-        }
-      }
-
-      // Read thread-persisted documentChatIds for follow-up messages
-      const dcStore = useDocumentChatStore.getState();
-      const documentChatIds = dcStore.getForThread(config.threadId);
-
-      const { fetch: configFetch, endpoints, contextProviders } = useChatConfigStore.getState();
-
-      // Surface-injected context (e.g. docs editor markdown + selected text).
-      // Keyed by threadId, so global chat threads and the docs thread coexist.
-      let injectedDocIds: string[] = [];
-      let injectedAttachmentContext: string | undefined;
-      let injectedCurrentDocument:
-        | {
-            id: string;
-            title?: string | null;
-            markdown: string;
-            selectionText?: string | null;
-          }
-        | undefined;
-      if (config.threadId) {
-        const provider = contextProviders.get(config.threadId);
-        if (provider) {
-          try {
-            const ctx = await provider();
-            if (ctx.documentChatIds?.length) injectedDocIds = ctx.documentChatIds;
-            if (ctx.currentDocument) {
-              const cd = ctx.currentDocument;
-              injectedCurrentDocument = {
-                id: cd.id,
-                title: cd.title ?? null,
-                markdown: truncateAttachmentContext(cd.markdown, 80_000) ?? cd.markdown,
-                selectionText: cd.selectionText ?? null,
-              };
-            }
-            const parts: string[] = [];
-            if (ctx.selectionText) parts.push(`## Auswahl:\n${ctx.selectionText}`);
-            if (ctx.attachmentContext) parts.push(ctx.attachmentContext);
-            const merged = parts.join('\n\n');
-            // Cap at 80k chars to keep request bodies bounded.
-            injectedAttachmentContext = merged
-              ? truncateAttachmentContext(merged, 80_000)
-              : undefined;
-          } catch (err) {
-            console.warn(
-              '[ChatAdapter] contextProvider threw, continuing without injected context',
-              err
-            );
-          }
-        }
-      }
-
-      const mergedDocChatIds = injectedDocIds.length
-        ? Array.from(new Set([...documentChatIds, ...injectedDocIds]))
-        : documentChatIds;
-
-      // Workplace flow seeds an Antrag/PM/Social text in agent store; pass it
-      // along on the FIRST request (no threadId yet) so the backend persists
-      // it as the seed assistant message of the brand-new thread.
-      const seededInitialAssistantMessage = !config.threadId
-        ? useAgentStore.getState().pendingInitialAssistantMessage || undefined
-        : undefined;
-
-      // Mode-aware endpoint selection (uses effectiveMode so search-routed agents
-      // dispatch to SearchGraph regardless of the store's threadMode).
-      const endpoint =
-        effectiveMode === 'search'
-          ? endpoints.searchStream
-          : effectiveMode === 'notebook'
-            ? endpoints.notebookStream
-            : endpoints.chatStream;
-
-      // Mode-aware request body
-      let requestBody: Record<string, unknown>;
-
-      if (effectiveMode === 'search') {
-        // Search mode: simple query + searchMode, no mentions/attachments
-        const lastUserText = formattedMessages
-          .filter((m) => m.role === 'user')
-          .pop()
-          ?.parts?.find((p: { type: string }) => p.type === 'text') as { text: string } | undefined;
-        requestBody = {
-          query: lastUserText?.text || '',
-          messages: formattedMessages,
-          threadId: config.threadId,
-          searchMode: config.searchMode || 'web',
-          agentId: config.agentId,
-        };
-      } else if (effectiveMode === 'notebook') {
-        // Notebook mode: query + collection scoping from selectedNotebookId
-        const lastUserText = formattedMessages
-          .filter((m) => m.role === 'user')
-          .pop()
-          ?.parts?.find((p: { type: string }) => p.type === 'text') as { text: string } | undefined;
-        requestBody = {
-          query: lastUserText?.text || '',
-          notebookId: config.selectedNotebookId || 'gruenerator-notebook',
-          threadId: config.threadId,
-        };
-      } else if (effectiveMode === 'eigener') {
-        // Eigener Chat mode: like chat but with custom prompt, no stale agentId
-        requestBody = {
-          messages: formattedMessages,
-          agentId: null,
-          threadId: config.threadId,
-          enabledTools: safeCustomEnabledTools
-            ? { ...config.enabledTools, ...safeCustomEnabledTools }
-            : config.enabledTools,
-          modelId: config.modelId,
-          attachments: extractedAttachments.length > 0 ? extractedAttachments : undefined,
-          notebookIds: notebookIds.length > 0 ? notebookIds : undefined,
-          forcedTools: forcedTools.length > 0 ? forcedTools : undefined,
-          documentIds: documentIds.length > 0 ? documentIds : undefined,
-          textIds: textIds.length > 0 ? textIds : undefined,
-          boardIds: boardIds.length > 0 ? boardIds : undefined,
-          docMentionIds: docMentionIds.length > 0 ? docMentionIds : undefined,
-          wolkeFiles: wolkeFiles.length > 0 ? wolkeFiles : undefined,
-          connectFiles: connectFiles.length > 0 ? connectFiles : undefined,
-          documentChatIds: mergedDocChatIds.length > 0 ? mergedDocChatIds : undefined,
-          documentChatMode: hasDocumentChat || mergedDocChatIds.length > 0 || undefined,
-          currentDocument: injectedCurrentDocument,
-          attachmentContext: injectedAttachmentContext,
-          defaultNotebookId: config.selectedNotebookId || undefined,
-          customSystemPrompt: config.customSystemPrompt || undefined,
-          roleName: config.customRoleName || undefined,
-          initialAssistantMessage: seededInitialAssistantMessage,
-          activeSkillMention: config.activeSkillMention || undefined,
-        };
-      } else {
-        // Chat mode: full request with mentions, attachments, tools
-        requestBody = {
-          messages: formattedMessages,
-          agentId: effectiveAgentId,
-          threadId: config.threadId,
-          enabledTools: safeCustomEnabledTools
-            ? { ...config.enabledTools, ...safeCustomEnabledTools }
-            : config.enabledTools,
-          modelId: config.modelId,
-          attachments: extractedAttachments.length > 0 ? extractedAttachments : undefined,
-          notebookIds: notebookIds.length > 0 ? notebookIds : undefined,
-          forcedTools: forcedTools.length > 0 ? forcedTools : undefined,
-          documentIds: documentIds.length > 0 ? documentIds : undefined,
-          textIds: textIds.length > 0 ? textIds : undefined,
-          boardIds: boardIds.length > 0 ? boardIds : undefined,
-          docMentionIds: docMentionIds.length > 0 ? docMentionIds : undefined,
-          wolkeFiles: wolkeFiles.length > 0 ? wolkeFiles : undefined,
-          connectFiles: connectFiles.length > 0 ? connectFiles : undefined,
-          documentChatIds: mergedDocChatIds.length > 0 ? mergedDocChatIds : undefined,
-          documentChatMode: hasDocumentChat || mergedDocChatIds.length > 0 || undefined,
-          currentDocument: injectedCurrentDocument,
-          attachmentContext: injectedAttachmentContext,
-          defaultNotebookId: config.selectedNotebookId || undefined,
-          customSystemPrompt: config.customSystemPrompt || undefined,
-          initialAssistantMessage: seededInitialAssistantMessage,
-          activeSkillMention: config.activeSkillMention || undefined,
-        };
-      }
-
-      let response: Response;
-      try {
-        response = await configFetch(endpoint, {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify(requestBody),
-          signal: abortSignal,
-        });
-      } catch (fetchError) {
-        if (abortSignal?.aborted) return;
-        yield { content: [{ type: 'text' as const, text: streamErrorMessage(fetchError) }] };
-        return;
-      }
-
-      if (!response.ok) {
-        yield { content: [{ type: 'text' as const, text: streamErrorMessage(null, response) }] };
-        return;
-      }
-
-      const streamOutcome: StreamOutcome = { interrupted: false, indexedDocumentIds: [] };
-      const resolvedAgentId = effectiveAgentId || config.agentId;
-      try {
-        yield* parseSSEStream(
-          response,
-          callbacks,
-          streamOutcome,
-          resolvedAgentId
-            ? { agentId: resolvedAgentId, agentMention: effectiveAgentMention }
-            : undefined
-        );
-      } catch (err) {
-        // Mid-stream connection drop (proxy timeout, mobile blip, worker recycle)
-        // surfaces as TypeError; treat it as a graceful end so it doesn't reach Sentry.
-        if (
-          err instanceof TypeError &&
-          /network error|failed to fetch|load failed|error in input stream/i.test(err.message)
-        ) {
-          return;
-        }
-        // Other mid-stream errors (e.g. backend SSE `error` event) — surface
-        // as an assistant message so the user sees what went wrong in-thread,
-        // not just in the dev console.
-        if (abortSignal?.aborted) return;
-        yield { content: [{ type: 'text' as const, text: streamErrorMessage(err) }] };
-        return;
-      }
-
-      // Persist server-indexed document IDs to thread for follow-up messages
-      if (streamOutcome.indexedDocumentIds.length > 0 && config.threadId) {
-        for (const docId of streamOutcome.indexedDocumentIds) {
-          dcStore.addToThread(config.threadId, docId);
-        }
-      }
-
-      if (streamOutcome.interrupted) {
-        interruptedThreadId = config.threadId;
-        lastInterruptedResult = streamOutcome.lastResult ?? null;
-      }
-    },
-  };
-}
